lyrics-qt: Split code into multiple reusable files
